Getting the working directory from pyLoad's passed arguments.
pyLoad passes some arguments about the file just being processed to the called script. This messed up this script, so I changed the behavior. Custom paths are now given by the --paths argument while the second of the automatically passed arguments is used as a default for the location to search for files.